+++ remote-backup/run.sh
@@ -13,9 +13,11 @@ ZIP_PASSWORD=$(bashio::config 'zip_password')
 KEEP_LOCAL_BACKUP=$(bashio::config 'keep_local_backup')
 
 RSYNC_ENABLED=$(bashio::config "rsync_enabled")
+RSYNC_VERBOSE=$(bashio::config "rsync_verbose")
 RSYNC_HOST=$(bashio::config "rsync_host")
 RSYNC_ROOTFOLDER=$(bashio::config "rsync_rootfolder")
 RSYNC_USER=$(bashio::config "rsync_user")
+RSYNC_EXCLUDE=$(bashio::config "rsync_exclude")
 RSYNC_PASSWORD=$(bashio::config "rsync_password")
 RCLONE_ENABLED=$(bashio::config "rclone_enabled")
 RCLONE_COPY=$(bashio::config "rclone_copy")
@@ -24,13 +26,17 @@ RCLONE_RESTORE=$(bashio::config "rclone_restore")
 RCLONE_REMOTE=$(bashio::config "rclone_remote")
 RCLONE_REMOTE_DIRECTORY=$(bashio::config "rclone_remote_directory")
 
+# define info messages
+function info { echo -e "\e[32m[info] $*\e[39m"; }
+function warn  { echo -e "\e[33m[warn] $*\e[39m"; }
+
 # create variables
 SSH_ID="/ssl/${SSH_KEY}"
 SSH_ID=$(echo -n ${SSH_ID})
 function add-ssh-key {
 
     if [ "$SSH_ENABLED" = true ] ; then
-        echo "Adding SSH key"
+        info "Adding SSH key"
         mkdir -p ~/.ssh
         cp ${SSH_ID} ${HOME}/.ssh/id_rsa
         chmod 600 "${HOME}/.ssh/id_rsa"
@@ -46,14 +52,15 @@ function add-ssh-key {
 
         chmod 600 "${HOME}/.ssh/config"
         chmod 644 "${HOME}/.ssh/id_rsa.pub"
+        info "SSH key added"
     fi    
 }
 
 function create-local-backup {
     name="${CUSTOM_PREFIX} $(date +'%Y-%m-%d %H-%M')"
-    echo "Creating local backup: \"${name}\""
+    warn "Creating local backup: \"${name}\""
     slug=$(ha backups new --raw-json --name="${name}" | jq --raw-output '.data.slug')
-    echo "Backup created: ${slug}"
+    info "Backup created: ${slug}"
 }
 
 function copy-backup-to-remote {
@@ -61,22 +68,27 @@ function copy-backup-to-remote {
     if [ "$SSH_ENABLED" = true ] ; then
         cd /backup/
         if [[ -z $ZIP_PASSWORD  ]]; then
-            echo "Copying ${slug}.tar to ${REMOTE_DIRECTORY} on ${SSH_HOST} using SCP"
+            warn "Copying ${slug}.tar to ${REMOTE_DIRECTORY} on ${SSH_HOST} using SCP"
             scp -F "${HOME}/.ssh/config" "${slug}.tar" remote:"${REMOTE_DIRECTORY}"
+            info "Backup copied to ${REMOTE_DIRECTORY}/${slug}.tar on ${SSH_HOST}"
         else
-            echo "Copying password-protected ${slug}.zip to ${REMOTE_DIRECTORY} on ${SSH_HOST} using SCP"
+            info "Copying password-protected ${slug}.zip to ${REMOTE_DIRECTORY} on ${SSH_HOST} using SCP"
             zip -P "$ZIP_PASSWORD" "${slug}.zip" "${slug}".tar
             scp -F "${HOME}/.ssh/config" "${slug}.zip" remote:"${REMOTE_DIRECTORY}" && rm "${slug}.zip"
+            info "Backup copied to ${REMOTE_DIRECTORY}/${slug}.zip on ${SSH_HOST}"
         fi
         if [ "$FRIENDLY_NAME" = true ] ; then
             if [[ -z $ZIP_PASSWORD  ]]; then
-                echo "Renaming ${slug}.tar to ${name}.tar"
+                warn "Renaming ${slug}.tar to ${name}.tar"
                 ssh remote "mv \"${REMOTE_DIRECTORY}/${slug}.tar\" \"${REMOTE_DIRECTORY}/${name}.tar\""
+                info "Backup renamed to ${REMOTE_DIRECTORY}/${name}.tar on ${SSH_HOST}"
             else
-                echo "Renaming ${slug}.zip to ${name}.zip"
+                info "Renaming ${slug}.zip to ${name}.zip"
                 ssh remote "mv \"${REMOTE_DIRECTORY}/${slug}.zip\" \"${REMOTE_DIRECTORY}/${name}.zip\""
+                info "Backup renamed to ${REMOTE_DIRECTORY}/${name}.zip on ${SSH_HOST}"
             fi
         fi
+    info "SCP complete"
     fi
 }
 
@@ -84,23 +96,59 @@ function rsync_folders {
 
     if [ "$RSYNC_ENABLED" = true ] ; then
         rsyncurl="$RSYNC_USER@$RSYNC_HOST:$RSYNC_ROOTFOLDER"
-        echo "[Info] trying to rsync ha folders to $rsyncurl"
-        echo ""
-        echo "[Info] /config"
-         sshpass -p $RSYNC_PASSWORD rsync -av --exclude '*.db-shm' --exclude '*.db-wal' --exclude '*.db' /config/ $rsyncurl/config/ --delete
-        echo ""
-        echo "[Info] /addons"
-         sshpass -p $RSYNC_PASSWORD rsync -av /addons/ $rsyncurl/addons/ --delete
-        echo ""
-        echo "[Info] /backup"
-         sshpass -p $RSYNC_PASSWORD rsync -av /backup/ $rsyncurl/backup/ --delete
-        echo ""
-        echo "[Info] /share"
-         sshpass -p $RSYNC_PASSWORD rsync -av /share/ $rsyncurl/share/ --delete
-        echo ""
-        echo "[Info] /ssl"
-         sshpass -p $RSYNC_PASSWORD rsync -av /ssl/ $rsyncurl/ssl/ --delete
-        echo "[Info] Finished rsync"
+        if [ "$RSYNC_VERBOSE" = true ] ; then
+            FLAGS='-av'
+        else
+            FLAGS='-a'
+        fi
+        if [ -z "$RSYNC_EXCLUDE" ]; then
+            warn "Syncing /config"
+             sshpass -p $RSYNC_PASSWORD rsync $FLAGS --exclude '*.db-shm' --exclude '*.db-wal' --exclude '*.db' /config/ $rsyncurl/config/ --delete
+            info "/config sync complete"
+            echo ""
+            warn "Syncing /addons"
+             sshpass -p $RSYNC_PASSWORD rsync $FLAGS /addons/ $rsyncurl/addons/ --delete
+            info "/addons sync complete"
+            echo ""
+            warn "Syncing /backup"
+             sshpass -p $RSYNC_PASSWORD rsync $FLAGS /backup/ $rsyncurl/backup/ --delete
+            info "/backup sync complete"
+            echo ""
+            warn "Syncing /share"
+             sshpass -p $RSYNC_PASSWORD rsync $FLAGS /share/ $rsyncurl/share/ --delete
+            info "/share sync complete"
+            echo ""
+            warn "Syncing /ssl"
+             sshpass -p $RSYNC_PASSWORD rsync $FLAGS /ssl/ $rsyncurl/ssl/ --delete
+            info "/ssl sync complete"
+            echo ""
+        else
+            echo $RSYNC_EXCLUDE | tr -s ", " "\n" > /tmp/rsync_exclude.txt
+            info "Files you excluded will be displayed below:"
+            cat /tmp/rsync_exclude.txt
+            info "Starting rsync"
+            warn "Syncing /config"
+             sshpass -p $RSYNC_PASSWORD rsync $FLAGS --exclude-from='/tmp/rsync_exclude.txt' --exclude '*.db-shm' --exclude '*.db-wal' --exclude '*.db' /config/ $rsyncurl/config/ --delete
+            info "/config sync complete"
+            echo ""
+            warn "Syncing /addons"
+             sshpass -p $RSYNC_PASSWORD rsync $FLAGS --exclude-from='/tmp/rsync_exclude.txt' /addons/ $rsyncurl/addons/ --delete
+            info "/addons sync complete"
+            echo ""
+            warn "Syncing /backup"
+             sshpass -p $RSYNC_PASSWORD rsync $FLAGS --exclude-from='/tmp/rsync_exclude.txt' /backup/ $rsyncurl/backup/ --delete
+            info "/backup sync complete"
+            echo ""
+            warn "Syncing /share"
+             sshpass -p $RSYNC_PASSWORD rsync $FLAGS --exclude-from='/tmp/rsync_exclude.txt' /share/ $rsyncurl/share/ --delete
+            info "/share sync complete"
+            echo ""
+            warn "Syncing /ssl"
+             sshpass -p $RSYNC_PASSWORD rsync $FLAGS --exclude-from='/tmp/rsync_exclude.txt' /ssl/ $rsyncurl/ssl/ --delete
+            info "/ssl sync complete"
+            echo ""
+        fi
+        info "Finished rsync"
     fi
 }
 
@@ -108,37 +156,43 @@ function rclone_backups {
     if [ "$RCLONE_ENABLED" = true ] ; then
         cd /backup/
         mkdir -p ~/.config/rclone/
-        cp -a /ssl/rclone.conf ~/.config/rclone/rclone.conf
+        cp $FLAGS /ssl/rclone.conf ~/.config/rclone/rclone.conf
         echo "Starting rclone"
         if [ "$RCLONE_COPY" = true ] ; then
             if [ "$FRIENDLY_NAME" = true ] ; then
                 if [[ -z $ZIP_PASSWORD  ]]; then
-                    echo "Copying ${slug}.tar to ${RCLONE_REMOTE_DIRECTORY}/${name}.tar"
+                    warn "Copying ${slug}.tar to ${RCLONE_REMOTE_DIRECTORY}/${name}.tar"
                     rclone copyto ${slug}.tar ${RCLONE_REMOTE}:${RCLONE_REMOTE_DIRECTORY}/"${name}".tar
+                    info "Finished rclone copy"
                 else
-                    echo "Copying ${slug}.zip to ${RCLONE_REMOTE_DIRECTORY}/${name}.zip"
+                    warn "Copying ${slug}.zip to ${RCLONE_REMOTE_DIRECTORY}/${name}.zip"
                     rclone copyto ${slug}.zip ${RCLONE_REMOTE}:${RCLONE_REMOTE_DIRECTORY}/"${name}".zip
+                    info "Finished rclone copy"
                 fi
             else
                 if [[ -z $ZIP_PASSWORD  ]]; then
-                    echo "Copying ${slug}.tar to ${RCLONE_REMOTE_DIRECTORY}/${slug}.tar"
+                    warn "Copying ${slug}.tar to ${RCLONE_REMOTE_DIRECTORY}/${slug}.tar"
                     rclone copy ${slug}.tar ${RCLONE_REMOTE}:${RCLONE_REMOTE_DIRECTORY}
+                    info "Finished rclone copy"
                 else
-                    echo "Copying ${slug}.zip to ${RCLONE_REMOTE_DIRECTORY}/${slug}.zip"
+                    warn "Copying ${slug}.zip to ${RCLONE_REMOTE_DIRECTORY}/${slug}.zip"
                     rclone copy ${slug}.zip ${RCLONE_REMOTE}:${RCLONE_REMOTE_DIRECTORY}
+                    info "Finished rclone copy"
                 fi
             fi
         fi
         if [ "$RCLONE_SYNC" = true ] ; then
-            echo "Syncing Backups"
+            warn "Syncing Backups"
             rclone sync . ${RCLONE_REMOTE}:${RCLONE_REMOTE_DIRECTORY}
+            info "Finished rclone sync"
         fi
         if [ "$RCLONE_RESTORE" = true ] ; then
             DATEFORMAT=$(date +%F)
             RESTORENAME="restore-${DATEFORMAT}"
             mkdir -p "${RESTORENAME}"
-            echo "Restoring Backups to ${RESTORENAME}"
+            warn "Restoring Backups to ${RESTORENAME}"
             rclone copyto ${RCLONE_REMOTE}:${RCLONE_REMOTE_DIRECTORY} ${RESTORENAME}/
+            info "Finished rclone restore"
         fi
     fi
 }
@@ -151,7 +205,7 @@ function delete-local-backup {
     if [[ ${KEEP_LOCAL_BACKUP} == "all" ]]; then
         :
     elif [[ -z ${KEEP_LOCAL_BACKUP} ]]; then
-        echo "Deleting local backup: ${slug}"
+        warn "Deleting local backup: ${slug}"
         ha backups remove "${slug}"
     else
 
@@ -160,8 +214,9 @@ function delete-local-backup {
 
         ha backups list --raw-json | jq -c .data.backups[] | while read backup; do
             if [[ $(echo ${backup} | jq .date | xargs date -D "%Y-%m-%dT%T" +%s --date ) -lt ${last_date_to_keep} ]]; then
-                echo "Deleting local backup: $(echo ${backup} | jq -r .slug)"
+                warn "Deleting local backup: $(echo ${backup} | jq -r .slug)"
                 ha backups remove "$(echo ${backup} | jq -r .slug)"
+                info "Finished deleting local backup: $(echo ${backup} | jq -r .slug)"
             fi
         done
 
@@ -175,5 +230,5 @@ rsync_folders
 rclone_backups
 delete-local-backup
 
-echo "Backup process done!"
+info "Backup process done!"
 exit 0